# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * AI Career Copilot API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UserProfile {
  id: number;
  clerkId: string;
  email: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  targetRole?: string | null;
  /** @nullable */
  experienceLevel?: string | null;
  /** @nullable */
  bio?: string | null;
  createdAt: string;
}

export interface UserProfileUpdate {
  name?: string;
  targetRole?: string;
  experienceLevel?: string;
  bio?: string;
}

export interface Resume {
  id: number;
  title: string;
  content: string;
  /** @nullable */
  atsScore?: number | null;
  isDefault: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface ResumeInput {
  title: string;
  content: string;
  isDefault?: boolean;
}

export interface ResumeAnalysis {
  id: number;
  resumeId: number;
  atsScore: number;
  overallFeedback: string;
  strengths: string[];
  weaknesses: string[];
  keywordsSuggested: string[];
  skillsFound: string[];
  skillsMissing: string[];
  formattingScore: number;
  contentScore: number;
  impactScore: number;
  createdAt: string;
}

export interface ImprovementSuggestion {
  section: string;
  issue: string;
  suggestion: string;
  priority: string;
}

export interface RewrittenSection {
  section: string;
  original: string;
  improved: string;
}

export interface ResumeImprovements {
  resumeId: number;
  suggestions: ImprovementSuggestion[];
  rewrittenSections: RewrittenSection[];
}

export interface JobDescription {
  id: number;
  title: string;
  /** @nullable */
  company?: string | null;
  description: string;
  requiredSkills: string[];
  createdAt: string;
}

export interface JobDescriptionInput {
  title: string;
  company?: string;
  description: string;
}

export interface JobMatchRequest {
  resumeId: number;
}

export interface JobMatch {
  id: number;
  jobId: number;
  resumeId: number;
  similarityScore: number;
  matchedSkills: string[];
  missingSkills: string[];
  recommendation: string;
  createdAt: string;
}

export interface RoadmapStep {
  id: number;
  roadmapId: number;
  title: string;
  description: string;
  /** @nullable */
  resourceUrl?: string | null;
  resourceType: string;
  /** @nullable */
  estimatedHours?: number | null;
  orderIndex: number;
  isCompleted: boolean;
}

export interface LearningRoadmap {
  id: number;
  title: string;
  targetRole: string;
  description: string;
  totalSteps: number;
  completedSteps: number;
  steps: RoadmapStep[];
  createdAt: string;
}

export interface RoadmapGenerateRequest {
  targetRole: string;
  currentSkills?: string[];
  missingSkills?: string[];
  /** @nullable */
  resumeId?: number | null;
  /** @nullable */
  jobId?: number | null;
}

export interface StepToggle {
  isCompleted: boolean;
}

export interface MockInterview {
  id: number;
  title: string;
  role: string;
  interviewType: string;
  status: string;
  totalQuestions: number;
  answeredQuestions: number;
  /** @nullable */
  overallScore?: number | null;
  /** @nullable */
  overallFeedback?: string | null;
  createdAt: string;
  /** @nullable */
  completedAt?: string | null;
}

export interface InterviewQuestion {
  id: number;
  interviewId: number;
  question: string;
  questionType: string;
  orderIndex: number;
  /** @nullable */
  userAnswer?: string | null;
  /** @nullable */
  feedback?: string | null;
  /** @nullable */
  score?: number | null;
  isAnswered: boolean;
  createdAt: string;
}

export interface MockInterviewDetail {
  id: number;
  title: string;
  role: string;
  interviewType: string;
  status: string;
  totalQuestions: number;
  answeredQuestions: number;
  /** @nullable */
  overallScore?: number | null;
  /** @nullable */
  overallFeedback?: string | null;
  questions: InterviewQuestion[];
  createdAt: string;
  /** @nullable */
  completedAt?: string | null;
}

export interface InterviewInput {
  role: string;
  interviewType: string;
  totalQuestions?: number;
  /** @nullable */
  resumeId?: number | null;
  /** @nullable */
  jobId?: number | null;
}

export interface AnswerInput {
  questionId: number;
  answer: string;
}

export interface AnswerFeedback {
  questionId: number;
  feedback: string;
  score: number;
  tips: string[];
  exampleAnswer: string;
}

export interface CoverLetter {
  id: number;
  title: string;
  content: string;
  /** @nullable */
  resumeId?: number | null;
  /** @nullable */
  jobId?: number | null;
  createdAt: string;
}

export interface CoverLetterRequest {
  resumeId: number;
  jobId: number;
  tone?: string;
}

export interface ChatConversation {
  id: number;
  title: string;
  messageCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface ChatMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface ChatConversationDetail {
  id: number;
  title: string;
  messages: ChatMessage[];
  createdAt: string;
  updatedAt: string;
}

export interface ChatConversationInput {
  title: string;
}

export interface ChatMessageInput {
  content: string;
}

export interface ScorePoint {
  date: string;
  score: number;
  label: string;
}

export interface DashboardStats {
  resumeCount: number;
  /** @nullable */
  avgAtsScore?: number | null;
  jobsTracked: number;
  interviewsCompleted: number;
  /** @nullable */
  avgInterviewScore?: number | null;
  roadmapsActive: number;
  stepsCompleted: number;
  coverLettersGenerated: number;
  topSkillGaps: string[];
  recentAtsScores: ScorePoint[];
}

export interface ActivityItem {
  id: string;
  type: string;
  title: string;
  description: string;
  createdAt: string;
}