# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Career Copilot API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cover-letters:
    get:
      operationId: listCoverLetters
      tags: [coverletters]
      summary: List all saved cover letters
      responses:
        "200":
          description: List of cover letters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CoverLetter"
    post:
      operationId: generateCoverLetter
      tags: [coverletters]
      summary: Generate a cover letter for a job application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CoverLetterRequest"
      responses:
        "201":
          description: Generated cover letter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoverLetter"
  /cover-letters/{id}:
    get:
      operationId: getCoverLetter
      tags: [coverletters]
      summary: Get a cover letter
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Cover letter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoverLetter"
        "404":
          description: Not found
    delete:
      operationId: deleteCoverLetter
      tags: [coverletters]
      summary: Delete a cover letter
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted

  # ── Chat (AI Career Chatbot) ───────────────────────────────────────────────
  /chat/conversations:
    get:
      operationId: listChatConversations
      tags: [chat]
      summary: List all chatbot conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatConversation"
    post:
      operationId: createChatConversation
      tags: [chat]
      summary: Start a new chatbot conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatConversationInput"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatConversation"
...
  /chat/conversations/{id}/messages:
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a message and receive a streaming AI response (SSE)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "200":
          description: SSE stream of AI response chunks
          content:
            text/event-stream: {}

  # ── Dashboard ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    CoverLetter:
      type: object
      properties:
        id:
          type: number
        title:
          type: string
        content:
          type: string
        resumeId:
          type: ["number", "null"]
        jobId:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - content
        - createdAt
    CoverLetterRequest:
      type: object
      properties:
        resumeId:
          type: number
        jobId:
          type: number
        tone:
          type: string
      required:
        - resumeId
        - jobId

    # ── Chat ───────────────────────────────────────────────────────────────
    ChatConversation:
      type: object
      properties:
        id:
          type: number
        title:
          type: string
        messageCount:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - messageCount
        - createdAt
        - updatedAt
...
    ChatConversationInput:
      type: object
      properties:
        title:
          type: string
      required:
        - title
    ChatMessageInput:
      type: object
      properties:
        content:
          type: string
      required:
        - content

    # ── Dashboard ──────────────────────────────────────────────────────────
...