# lib/api-zod/src/generated/types/coverLetterRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * AI Career Copilot API
 * OpenAPI spec version: 0.1.0
 */

export interface CoverLetterRequest {
  resumeId: number;
  jobId: number;
  tone?: string;
}